Allowing for passing in a pre-existing binder to LambdaCompiler.Compile
This allows you to “cache” the binder for multiple snippets bound to
the same context type, and clone it before you use it to evaluate your
code. This should save significant amount of resources, since the
initial creation of a binder implies some initial reflection to
retrieve the methods with the Bind attribute and create delegates out
of these, to make them accessible for your Lizzie code.